refactor(ui): extract accumulated report spec and shared actions
Extract AccumulatedOverviewPanel.jsx's report-spec-registration block
into accumulatedReportSpecs.jsx (mirrors runReportSpecs.jsx), and
useSharedProjects.js's connect/pull ref-guarded closures into
useSharedActions.js (mirrors usePublishTrigger's in-flight-ref idiom).
Both extractions are verbatim relocations; race guards intact.
